feature: add up coming section to home screen
