+++ presentation/src/main/java/com/london/presentation/screen/home/HomeScreen.kt
@@ -1,34 +1,98 @@
 package com.london.presentation.screen.home
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawing
-import androidx.compose.foundation.layout.windowInsetsPadding
+import androidx.compose.foundation.layout.requiredWidth
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.GridItemSpan
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.london.designsystem.component.Scaffold
+import com.london.designsystem.component.HomeCard
+import com.london.designsystem.component.NovixChip
 import com.london.designsystem.component.Text
 import com.london.designsystem.theme.NovixTheme
+import com.london.designsystem.theme.ThemePreviews
+import com.london.presentation.R
 
 @Composable
-fun HomeScreen(modifier: Modifier = Modifier) {
-    Scaffold(containerColor = NovixTheme.colors.surface) { innerPadding ->
-        Box(
-            modifier = modifier
-                .fillMaxSize()
-                .padding(innerPadding)
-                .padding(16.dp)
-                .windowInsetsPadding(WindowInsets.safeDrawing),
-            contentAlignment = Alignment.Center
-        ) {
+fun HomeScreen() {
+        Content(
+            onMovieClick = { },
+            onGenreClick = { },
+            modifier = Modifier.padding(top = 100.dp).fillMaxSize()
+        )
+}
+
+@Composable
+private fun Content(
+    onMovieClick: (movieId: Int) -> Unit,
+    onGenreClick: (genreId: Int) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val density = LocalDensity.current
+    val screenWidth = with(density) {
+        LocalConfiguration.current.screenWidthDp.dp
+    }
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(
+            top = 12.dp,
+            bottom = 16.dp,
+            start = 16.dp,
+            end = 16.dp
+        ),
+        horizontalArrangement = Arrangement.spacedBy(12.dp),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+        modifier = modifier
+            .background(color = NovixTheme.colors.surface)
+    ) {
+        item(span = { GridItemSpan(maxLineSpan) }) {
             Text(
-                text = "Home Screen",
-                style = NovixTheme.typography.headline.medium
+                text = stringResource(R.string.upcoming),
+                style = NovixTheme.typography.headline.small,
+                color = NovixTheme.colors.title
             )
         }
+        item(span = { GridItemSpan(maxLineSpan) }) {
+            LazyRow(
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+                contentPadding = PaddingValues(horizontal = 16.dp),
+                modifier = Modifier.requiredWidth(screenWidth)
+            ) {
+                items(10) {
+                    NovixChip(
+                        text = "Adventure",//TODO()
+                        isSelected = true,
+                        onClick = { onGenreClick(it) }
+                    )
+                }
+            }
+        }
+        items(10) {
+            HomeCard(
+                imageUrl = "",//TODO()
+                isSaved = false,
+                onSaveClick = { },//TODO()
+                modifier = Modifier.clickable { onMovieClick(it) })
+        }
     }
+}
+
+@ThemePreviews
+@Composable
+private fun Preview() {
+    Content(
+        onMovieClick = { },
+        onGenreClick = { },
+    )
 }
