+++ presentation/src/main/res/values-ar/strings.xml
@@ -64,5 +64,6 @@
     <string name="actor_photos">صور الممثلين</string>"
     <string name="author_profile">صورة الممثل</string>
     <string name="tv_show_image">صور مسلسلات</string>
+    <string name="upcoming">القادم</string>">
 
 </resources>
