+++ presentation/src/main/res/values/strings.xml
@@ -64,4 +64,5 @@
     <string name="actor_photos">Actor Photos</string>
     <string name="author_profile">Author profile</string>
     <string name="tv_show_image">TV Show Image</string>
+    <string name="upcoming">Upcoming</string>">
 </resources>
